work/streamlit/app/requirements.txt
streamlit
wordcloud
matplotlib
kafka-python
pandas
pymongo
